Moving runtime.exs into a module
